Updates to Unix tools for bioinformatics
Note that in the present state I am unable to export the .org mode to .tex. The .tex
file modified here is the output of the previous .org file which I had to modify in
order to run pdflatex, as it did not like the parameter specifications for the figures
produced by .org.